+++ js/views/games/findType.js
@@ -1,10 +1,10 @@
-import AbstractView from '../view.js';
+import AbstractView from '../abstractView.js';
 import header from '../../blocks/header.js';
 import levelStats from '../../blocks/levelStats.js';
 import {changeAspectRatioOnLoad} from '../../utils/utils.js';
 import {initialState, stats} from '../../data/data.js';
 
-export default class GameTwoView extends AbstractView {
+export default class GameThreeView extends AbstractView {
   constructor(question) {
     super();
     this.question = question;
@@ -14,19 +14,11 @@ export default class GameTwoView extends AbstractView {
       ${header(initialState)}
         <div class="game">
           <p class="game__task">${this.question.question}</p>
-          <form class="game__content game__content--wide">
+          <form class="game__content game__content--triple">
             ${this.question.answers.map((answer, i) =>
               `<div class="game__option">
                 <img src="${answer.image}" alt="Option ${i + 1}">
-                <label class="game__answer game__answer--photo">
-                  <input name="question1" type="radio" value="photo">
-                  <span>Фото</span>
-                </label>
-                <label class="game__answer game__answer--paint">
-                  <input name="question1" type="radio" value="paint">
-                  <span>Рисунок</span>
-                </label>
-                </div>`).join(``)
+              </div>`).join(``)
             }
           </form>
           <div class="stats">
@@ -39,23 +31,27 @@ export default class GameTwoView extends AbstractView {
   }
 
   bind() {
+    const gameOptions = this.element.querySelectorAll(`.game__option`);
+    for (const option of gameOptions) {
+      option.addEventListener(`click`, () => {
+        this.onChangeScreen();
+      });
+    }
+
     const backButton = this.element.querySelector(`.header__back`);
     backButton.addEventListener(`click`, () => {
       this.onBackButtonClick();
     });
 
-    const answerInputs = this.element.querySelectorAll(`.game__answer input`);
-    for (const answer of answerInputs) {
-      answer.addEventListener(`change`, () => {
-        this.onChangeScreen();
-      });
-    }
-
     const images = this.element.querySelectorAll(`.game__option > img`);
     changeAspectRatioOnLoad(images);
   }
 
-  onBackButtonClick() {}
+  onBackButtonClick() {
+    throw new Error(`Not implemented onBackButtonClick`);
+  }
 
-  onChangeScreen() {}
+  onChangeScreen() {
+    throw new Error(`Not implemented onChangeScreen`);
+  }
 }
