+++ js/views/intro/intro.js
@@ -1,6 +1,6 @@
 import showGreeting from '../greeting/greeting.js';
 import {setScreen} from '../../utils/utils.js';
-import IntroView from './intro-view.js';
+import IntroView from './introView.js';
 
 export default () => {
   const intro = new IntroView();
