+++ js/views/rules/rulesView.js
@@ -1,4 +1,4 @@
-import AbstractView from '../view.js';
+import AbstractView from '../abstractView.js';
 import header from '../../blocks/header.js';
 
 export default class RulesView extends AbstractView {
