+++ js/views/stats/stats.js
@@ -1,6 +1,6 @@
 import {setScreen} from '../../utils/utils.js';
 import greetingScreen from '../greeting/greeting.js';
-import StatsView from './stats-view.js';
+import StatsView from './statsView.js';
 
 export default () => {
   const statsScreen = new StatsView();
